backend/src/test/java/ma/zs/zyn/unit/ws.facade/admin/template/TechnologyRestAdminTest.java
package ma.zs.zyn.unit.ws.facade.admin.template;

import ma.zs.zyn.bean.core.template.Technology;
import ma.zs.zyn.service.impl.admin.template.TechnologyAdminServiceImpl;
import ma.zs.zyn.ws.facade.admin.template.TechnologyRestAdmin;
import ma.zs.zyn.ws.converter.template.TechnologyConverter;
import ma.zs.zyn.ws.dto.template.TechnologyDto;
import org.aspectj.lang.annotation.Before;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;

import java.util.Collections;
import java.util.List;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.mockito.Mockito.when;

@ExtendWith(MockitoExtension.class)
public class TechnologyRestAdminTest {

    private MockMvc mockMvc;

    @Mock
    private TechnologyAdminServiceImpl service;
    @Mock
    private TechnologyConverter converter;

    @InjectMocks
    private TechnologyRestAdmin controller;

    @Before("")
    public void setUp() {
        MockitoAnnotations.initMocks(this);
        this.mockMvc = MockMvcBuilders.standaloneSetup(controller).build();
    }


    @Test
    public void itShouldFindAllTechnologyTest() throws Exception {
        // Mock the service to return an empty list
        when(service.findAll()).thenReturn(Collections.emptyList());
        when(converter.toDto(Collections.emptyList())).thenReturn(Collections.emptyList());

        // Call the controller method
        ResponseEntity<List<TechnologyDto>> result = controller.findAll();

        // Verify the result
        assertEquals(HttpStatus.NO_CONTENT, result.getStatusCode());

        // Response body should be empty list
        List<TechnologyDto> responseBody = result.getBody();
        assertNotNull(responseBody);
        assertEquals(0, responseBody.size());
    }

    @Test
    public void itShouldSaveTechnologyTest() throws Exception {
        // Mock data
        TechnologyDto requestDto = new TechnologyDto();
        Technology entity = new Technology();
        Technology saved = new Technology();
        TechnologyDto savedDto = new TechnologyDto();

        // Mock the converter to return the technology object when converting from DTO
        when(converter.toItem(requestDto)).thenReturn(entity);

        // Mock the service to return the saved client
        when(service.create(entity)).thenReturn(saved);

        // Mock the converter to return the saved technology DTO
        when(converter.toDto(saved)).thenReturn(savedDto);

        // Call the controller method
        ResponseEntity<TechnologyDto> result = controller.save(requestDto);

        // Verify the result
        assertEquals(HttpStatus.CREATED, result.getStatusCode());

        // Verify the response body
        TechnologyDto responseBody = result.getBody();
        assertNotNull(responseBody);

        // Add assertions to compare the response body with the saved technology DTO
        assertEquals(savedDto.getCode(), responseBody.getCode());
        assertEquals(savedDto.getLibelle(), responseBody.getLibelle());
        assertEquals(savedDto.getLogo(), responseBody.getLogo());
    }

}
